+++ py7zr/archiveinfo.py
@@ -26,11 +26,12 @@ import functools
 import io
 import os
 import struct
-from array import array
 from binascii import unhexlify
 from functools import reduce
+from io import BytesIO
 from operator import and_, or_
 from struct import pack, unpack
+from typing import Any, BinaryIO, Dict, List, Optional, Tuple
 
 from py7zr.compression import SevenZipCompressor, SevenZipDecompressor
 from py7zr.exceptions import Bad7zFile, UnsupportedCompressionMethodError
@@ -41,25 +42,25 @@ from py7zr.properties import (MAGIC_7Z, CompressionMethod, Configuration,
 MAX_LENGTH = 65536
 
 
-def read_crcs(file, count):
+def read_crcs(file: BinaryIO, count: int) -> List[int]:
     data = file.read(4 * count)
     return [unpack('<L', data[i * 4:i * 4 + 4])[0] for i in range(count)]
 
 
-def write_crcs(file, crcs):
+def write_crcs(file: BinaryIO, crcs):
     for crc in crcs:
         write_uint32(file, crc)
 
 
-def read_bytes(file, length):
+def read_bytes(file: BinaryIO, length: int) -> Tuple[bytes, ...]:
     return unpack(b'B' * length, file.read(length))
 
 
-def read_byte(file):
+def read_byte(file: BinaryIO) -> int:
     return ord(file.read(1))
 
 
-def write_bytes(file, data):
+def write_bytes(file: BinaryIO, data):
     if isinstance(data, bytes):
         return file.write(data)
     elif isinstance(data, bytearray):
@@ -69,52 +70,59 @@ def write_bytes(file, data):
         raise
 
 
-def write_byte(file, data):
+def write_byte(file: BinaryIO, data):
     assert len(data) == 1
     return write_bytes(file, data)
 
 
-def read_real_uint64(file):
+def read_real_uint64(file: BinaryIO) -> Tuple[int, bytes]:
     """read 8 bytes, return unpacked value as a little endian unsigned long long, and raw data."""
     res = file.read(8)
     a = unpack('<Q', res)[0]
     return a, res
 
 
-def read_uint32(file):
+def read_uint32(file: BinaryIO) -> Tuple[int, bytes]:
     """read 4 bytes, return unpacked value as a little endian unsigned long, and raw data."""
     res = file.read(4)
     a = unpack('<L', res)[0]
     return a, res
 
 
-def write_uint32(file, value):
+def write_uint32(file: BinaryIO, value):
     """write uint32 value in 4 bytes."""
     b = pack('<L', value)
     file.write(b)
 
 
-def read_uint64(file):
+def read_uint64(file: BinaryIO) -> int:
+    """read UINT64, definition show in write_uint64()"""
     b = ord(file.read(1))
-    mask = 0x80
     if b == 255:
         return read_real_uint64(file)[0]
-    for i in range(8):
-        if b & mask == 0:
-            bytes = array('B', file.read(i))
-            bytes.reverse()
-            value = (bytes and reduce(lambda x, y: x << 8 | y, bytes)) or 0
-            highpart = b & (mask - 1)
-            return value + (highpart << (i * 8))
+    blen = [(0b01111111, 0), (0b10111111, 1), (0b11011111, 2), (0b11101111, 3),
+            (0b11110111, 4), (0b11111011, 5), (0b11111101, 6), (0b11111110, 7)]
+    mask = 0x80
+    vlen = 8
+    for v, l in blen:
+        if b <= v:
+            vlen = l
+            break
         mask >>= 1
+    if vlen == 0:
+        return b & (mask - 1)
+    val = file.read(vlen)
+    value = int.from_bytes(val, byteorder='little')
+    highpart = b & (mask - 1)
+    return value + (highpart << (vlen * 8))
 
 
-def write_real_uint64(file, value):
+def write_real_uint64(file: BinaryIO, value: int):
     """write 8 bytes, as an unsigned long long."""
     file.write(pack('<Q', value))
 
 
-def write_uint64(file, value):
+def write_uint64(file: BinaryIO, value: int):
     """
     UINT64 means real UINT64 encoded with the following scheme:
 
@@ -143,7 +151,7 @@ def write_uint64(file, value):
     file.write(ba)
 
 
-def read_boolean(file, count, checkall=False):
+def read_boolean(file: BinaryIO, count: int, checkall: int = False) -> List[bool]:
     if checkall:
         all_defined = file.read(1)
         if all_defined != unhexlify('00'):
@@ -160,7 +168,7 @@ def read_boolean(file, count, checkall=False):
     return result
 
 
-def write_boolean(file, booleans, all_defined=False):
+def write_boolean(file: BinaryIO, booleans: List[bool], all_defined: bool = False):
     if all_defined and reduce(and_, booleans):
         file.write(b'\x01')
         return
@@ -180,7 +188,7 @@ def write_boolean(file, booleans, all_defined=False):
         file.write(pack('B', o))
 
 
-def read_utf16(file):
+def read_utf16(file: BinaryIO) -> str:
     """read a utf-16 string from file"""
     val = ''
     for _ in range(MAX_LENGTH):
@@ -191,7 +199,7 @@ def read_utf16(file):
     return val
 
 
-def write_utf16(file, val):
+def write_utf16(file: BinaryIO, val: str):
     """write a utf-16 string to file"""
     file.write(val.encode('utf-16'))
     file.write(unhexlify(('0000')))
@@ -234,17 +242,17 @@ class PackInfo:
 
     __slots__ = ['packpos', 'numstreams', 'packsizes', 'packpositions', 'crcs']
 
-    def __init__(self):
-        self.packpos = None
-        self.numstreams = None
-        self.packsizes = []
-        self.crcs = None
+    def __init__(self) -> None:
+        self.packpos = 0  # type: int
+        self.numstreams = 0  # type: int
+        self.packsizes = []  # type: List[int]
+        self.crcs = None  # type: Optional[List[int]]
 
     @classmethod
-    def retrieve(cls, file):
+    def retrieve(cls, file: BinaryIO):
         return cls()._read(file)
 
-    def _read(self, file):
+    def _read(self, file: BinaryIO):
         self.packpos = read_uint64(file)
         self.numstreams = read_uint64(file)
         pid = file.read(1)
@@ -256,10 +264,10 @@ class PackInfo:
                 pid = file.read(1)
         if pid != Property.END:
             raise Bad7zFile('end id expected but %s found' % repr(pid))
-        self.packpositions = [sum(self.packsizes[:i]) for i in range(self.numstreams)]
+        self.packpositions = [sum(self.packsizes[:i]) for i in range(self.numstreams)]  # type: List[int]
         return self
 
-    def write(self, file):
+    def write(self, file: BinaryIO):
         assert self.packpos is not None
         numstreams = len(self.packsizes)
         assert self.crcs is None or len(self.crcs) == numstreams
@@ -288,30 +296,30 @@ class Folder:
     __slots__ = ['unpacksizes', 'solid', 'num_coders', 'coders', 'digestdefined', 'totalin', 'totalout',
                  'bindpairs', 'packed_indices', 'queue', 'crc', 'decompressor', 'compressor', 'files']
 
-    def __init__(self):
-        self.unpacksizes = None
-        self.coders = []
-        self.bindpairs = []
-        self.packed_indices = []
+    def __init__(self) -> None:
+        self.unpacksizes = None  # type: Optional[List[int]]
+        self.coders = []  # type: List[Any]
+        self.bindpairs = []  # type: List[Any]
+        self.packed_indices = []  # type: List[Any]
         # calculated values
-        self.totalin = 0
-        self.totalout = 0
+        self.totalin = 0  # type: int
+        self.totalout = 0  # type: int
         # internal values
-        self.solid = False
-        self.digestdefined = False
-        self.crc = None
+        self.solid = False  # type: bool
+        self.digestdefined = False  # type: bool
+        self.crc = None  # type: Optional[int]
         # compress/decompress objects
-        self.decompressor = None
-        self.compressor = None
+        self.decompressor = None  # type: Optional[SevenZipDecompressor]
+        self.compressor = None  # type: Optional[SevenZipCompressor]
         self.files = None
 
     @classmethod
-    def retrieve(cls, file):
+    def retrieve(cls, file: BinaryIO):
         obj = cls()
         obj._read(file)
         return obj
 
-    def _read(self, file):
+    def _read(self, file: BinaryIO) -> None:
         num_coders = read_uint64(file)
         for i in range(num_coders):
             while True:
@@ -320,7 +328,7 @@ class Folder:
                 iscomplex = b & 0x10 == 0x10
                 hasattributes = b & 0x20 == 0x20
                 last_alternative = b & 0x80 == 0
-                c = {'method': file.read(methodsize)}
+                c = {'method': file.read(methodsize)}  # type: Dict[str, Any]
                 if iscomplex:
                     c['numinstreams'] = read_uint64(file)
                     c['numoutstreams'] = read_uint64(file)
@@ -347,7 +355,7 @@ class Folder:
             for i in range(num_packedstreams):
                 self.packed_indices.append(read_uint64(file))
 
-    def write(self, file):
+    def write(self, file: BinaryIO):
         num_coders = len(self.coders)
         assert num_coders > 0
         write_uint64(file, num_coders)
@@ -358,20 +366,15 @@ class Folder:
             numoutstreams = c['numoutstreams']
             iscomplex = 0x00 if numinstreams == 1 and numoutstreams == 1 else 0x10
             last_alternative = 0x80 if i < num_coders - 1 else 0x00
-            if c['properties'] is not None:
-                hasattributes = 0x20
-                properties = c['properties']
-            else:
-                hasattributes = 0x00
-                properties = None
+            hasattributes = 0x20 if c['properties'] is not None else 0x00
             write_byte(file, struct.pack('B', method_size & 0xf | iscomplex | hasattributes | last_alternative))
             write_bytes(file, method)
-            if iscomplex:
+            if iscomplex == 0x10:
                 write_uint64(file, numinstreams)
                 write_uint64(file, numoutstreams)
-            if properties is not None:
-                write_uint64(file, len(properties))
-                write_bytes(file, properties)
+            if c['properties'] is not None:
+                write_uint64(file, len(c['properties']))
+                write_bytes(file, c['properties'])
         num_bindpairs = self.totalout - 1
         assert len(self.bindpairs) == num_bindpairs
         num_packedstreams = self.totalin - num_bindpairs
@@ -382,7 +385,7 @@ class Folder:
             for pi in self.packed_indices:
                 write_uint64(file, pi)
 
-    def get_decompressor(self, size):
+    def get_decompressor(self, size: int) -> SevenZipDecompressor:
         if self.decompressor is not None:
             return self.decompressor
         else:
@@ -390,20 +393,24 @@ class Folder:
                 self.decompressor = SevenZipDecompressor(self.coders, size, self.crc)
             except Exception as e:
                 raise e
-            return self.decompressor
+            if self.decompressor is not None:
+                return self.decompressor
+            else:
+                raise
 
-    def get_compressor(self):
+    def get_compressor(self) -> SevenZipCompressor:
         if self.compressor is not None:
             return self.compressor
         else:
             try:
-                self.compressor = SevenZipCompressor()  # FIXME: set filters
+                # FIXME: set filters
+                self.compressor = SevenZipCompressor()
                 self.coders = self.compressor.coders
+                return self.compressor
             except Exception as e:
                 raise e
-            return self.compressor
 
-    def get_unpack_size(self):
+    def get_unpack_size(self) -> int:
         if self.unpacksizes is None:
             return 0
         for i in range(len(self.unpacksizes) - 1, -1, -1):
@@ -411,19 +418,19 @@ class Folder:
                 return self.unpacksizes[i]
         raise TypeError('not found')
 
-    def _find_in_bin_pair(self, index):
+    def _find_in_bin_pair(self, index: int) -> int:
         for idx, (a, b) in enumerate(self.bindpairs):
             if a == index:
                 return idx
         return -1
 
-    def _find_out_bin_pair(self, index):
+    def _find_out_bin_pair(self, index: int) -> int:
         for idx, (a, b) in enumerate(self.bindpairs):
             if b == index:
                 return idx
         return -1
 
-    def is_encrypted(self):
+    def is_encrypted(self) -> bool:
         return CompressionMethod.CRYPT_AES256_SHA256 in [x['method'] for x in self.coders]
 
 
@@ -433,7 +440,7 @@ class UnpackInfo:
     __slots__ = ['numfolders', 'folders', 'datastreamidx']
 
     @classmethod
-    def retrieve(cls, file):
+    def retrieve(cls, file: BinaryIO):
         obj = cls()
         obj._read(file)
         return obj
@@ -443,7 +450,7 @@ class UnpackInfo:
         self.folders = []
         self.datastreamidx = None
 
-    def _read(self, file):
+    def _read(self, file: BinaryIO):
         pid = file.read(1)
         if pid != Property.FOLDER:
             raise Bad7zFile('folder id expected but %s found' % repr(pid))
@@ -460,7 +467,7 @@ class UnpackInfo:
             file.seek(current_pos, 0)
         self._retrieve_coders_info(file)
 
-    def _retrieve_coders_info(self, file):
+    def _retrieve_coders_info(self, file: BinaryIO):
         pid = file.read(1)
         if pid != Property.CODERS_UNPACK_SIZE:
             raise Bad7zFile('coders unpack size id expected but %s found' % repr(pid))
@@ -468,7 +475,7 @@ class UnpackInfo:
             folder.unpacksizes = [read_uint64(file) for _ in range(folder.totalout)]
         pid = file.read(1)
         if pid == Property.CRC:
-            defined = read_boolean(file, self.numfolders, checkall=1)
+            defined = read_boolean(file, self.numfolders, checkall=True)
             crcs = read_crcs(file, self.numfolders)
             for idx, folder in enumerate(self.folders):
                 folder.digestdefined = defined[idx]
@@ -477,7 +484,7 @@ class UnpackInfo:
         if pid != Property.END:
             raise Bad7zFile('end id expected but %s found' % repr(pid))
 
-    def write(self, file):
+    def write(self, file: BinaryIO):
         file.write(Property.UNPACK_INFO)
         file.write(Property.FOLDER)
         write_uint64(file, self.numfolders)
@@ -501,15 +508,18 @@ class UnpackInfo:
 class SubstreamsInfo:
     """ defines the substreams of a folder """
 
+    def __init__(self):
+        self.digests = []  # type: List[int]
+        self.digestsdefined = []  # type: List[bool]
+        self.unpacksizes = None  # type: Optional[List[int]]
+
     @classmethod
-    def retrieve(cls, file, numfolders, folders):
+    def retrieve(cls, file: BinaryIO, numfolders: int, folders: List[Folder]):
         obj = cls()
         obj._read(file, numfolders, folders)
         return obj
 
-    def _read(self, file, numfolders, folders):
-        self.digests = []
-        self.digestsdefined = []
+    def _read(self, file: BinaryIO, numfolders: int, folders: List[Folder]):
         pid = file.read(1)
         if pid == Property.NUM_UNPACK_STREAM:
             self.num_unpackstreams_folders = [read_uint64(file) for _ in range(numfolders)]
@@ -519,7 +529,7 @@ class SubstreamsInfo:
         if pid == Property.SIZE:
             self.unpacksizes = []
             for i in range(len(self.num_unpackstreams_folders)):
-                totalsize = 0
+                totalsize = 0  # type: int
                 for j in range(1, self.num_unpackstreams_folders[i]):
                     size = read_uint64(file)
                     self.unpacksizes.append(size)
@@ -540,7 +550,7 @@ class SubstreamsInfo:
             for i in range(numfolders):
                 folder = folders[i]
                 numsubstreams = self.num_unpackstreams_folders[i]
-                if numsubstreams == 1 and folder.digestdefined:
+                if numsubstreams == 1 and folder.digestdefined and folder.crc is not None:
                     self.digestsdefined.append(True)
                     self.digests.append(folder.crc)
                 else:
@@ -555,10 +565,12 @@ class SubstreamsInfo:
             self.digestsdefined = [False] * num_digests_total
             self.digests = [0] * num_digests_total
 
-    def write(self, file, numfolders):
+    def write(self, file: BinaryIO, numfolders: int):
         if self.num_unpackstreams_folders is None or len(self.num_unpackstreams_folders) == 0:
             # nothing to write
             return
+        if self.unpacksizes is None:
+            raise ValueError
         write_byte(file, Property.SUBSTREAMS_INFO)
         if not functools.reduce(lambda x, y: x and (y == 1), self.num_unpackstreams_folders, True):
             write_byte(file, Property.NUM_UNPACK_STREAM)
@@ -585,17 +597,17 @@ class StreamsInfo:
     __slots__ = ['packinfo', 'unpackinfo', 'substreamsinfo']
 
     def __init__(self):
-        self.packinfo = None
-        self.unpackinfo = None
-        self.substreamsinfo = None
+        self.packinfo = None  # type: PackInfo
+        self.unpackinfo = None  # type: UnpackInfo
+        self.substreamsinfo = None  # type: SubstreamsInfo
 
     @classmethod
-    def retrieve(cls, file):
+    def retrieve(cls, file: BinaryIO):
         obj = cls()
         obj.read(file)
         return obj
 
-    def read(self, file):
+    def read(self, file: BinaryIO) -> None:
         pid = file.read(1)
         if pid == Property.PACK_INFO:
             self.packinfo = PackInfo.retrieve(file)
@@ -609,7 +621,7 @@ class StreamsInfo:
         if pid != Property.END:
             raise Bad7zFile('end id expected but %s found' % repr(pid))
 
-    def write(self, file):
+    def write(self, file: BinaryIO):
         if self.packinfo is not None:
             write_byte(file, Property.PACK_INFO)
             self.packinfo.write(file)
@@ -633,12 +645,12 @@ class FilesInfo:
         self.dataindex = None
 
     @classmethod
-    def retrieve(cls, file):
+    def retrieve(cls, file: BinaryIO):
         obj = cls()
         obj._read(file)
         return obj
 
-    def _read(self, fp):
+    def _read(self, fp: BinaryIO):
         self.numfiles = read_uint64(fp)
         self.files = [{'emptystream': False} for _ in range(self.numfiles)]
         numemptystreams = 0
@@ -646,16 +658,16 @@ class FilesInfo:
             typ = read_uint64(fp)
             if typ > 255:
                 raise Bad7zFile('invalid type, must be below 256, is %d' % typ)
-            typ = struct.pack('B', typ)
-            if typ == Property.END:
+            prop = struct.pack(b'B', typ)
+            if prop == Property.END:
                 break
             size = read_uint64(fp)
-            if typ == Property.DUMMY:
+            if prop == Property.DUMMY:
                 # Added by newer versions of 7z to adjust padding.
                 fp.seek(size, os.SEEK_CUR)
                 continue
             buffer = io.BytesIO(fp.read(size))
-            if typ == Property.EMPTY_STREAM:
+            if prop == Property.EMPTY_STREAM:
                 isempty = read_boolean(buffer, self.numfiles)
                 list(map(lambda x, y: x.update({'emptystream': y}), self.files, isempty))
                 for x in isempty:
@@ -663,11 +675,11 @@ class FilesInfo:
                         numemptystreams += 1
                 self.emptyfiles = [False] * numemptystreams
                 self.antifiles = [False] * numemptystreams
-            elif typ == Property.EMPTY_FILE:
+            elif prop == Property.EMPTY_FILE:
                 self.emptyfiles = read_boolean(buffer, numemptystreams)
-            elif typ == Property.ANTI:
+            elif prop == Property.ANTI:
                 self.antifiles = read_boolean(buffer, numemptystreams)
-            elif typ == Property.NAME:
+            elif prop == Property.NAME:
                 external = buffer.read(1)
                 if external == b'\x00':
                     self._read_name(buffer)
@@ -678,13 +690,13 @@ class FilesInfo:
                     fp.seek(self.dataindex, 0)
                     self._read_name(buffer)
                     fp.seek(current_pos, 0)
-            elif typ == Property.CREATION_TIME:
+            elif prop == Property.CREATION_TIME:
                 self._readTimes(buffer, self.files, 'creationtime')
-            elif typ == Property.LAST_ACCESS_TIME:
+            elif prop == Property.LAST_ACCESS_TIME:
                 self._readTimes(buffer, self.files, 'lastaccesstime')
-            elif typ == Property.LAST_WRITE_TIME:
+            elif prop == Property.LAST_WRITE_TIME:
                 self._readTimes(buffer, self.files, 'lastwritetime')
-            elif typ == Property.ATTRIBUTES:
+            elif prop == Property.ATTRIBUTES:
                 defined = read_boolean(buffer, self.numfiles, checkall=1)
                 external = buffer.read(1)
                 if external == b'\x00':
@@ -697,20 +709,20 @@ class FilesInfo:
                     self._read_attributes(fp, defined)
                     fp.seek(current_pos, 0)
             else:
-                raise Bad7zFile('invalid type %r' % (typ))
+                raise Bad7zFile('invalid type %r' % (prop))
 
-    def _read_name(self, buffer):
+    def _read_name(self, buffer: BinaryIO) -> None:
         for f in self.files:
             f['filename'] = read_utf16(buffer)
 
-    def _read_attributes(self, buffer, defined):
+    def _read_attributes(self, buffer: BinaryIO, defined: List[bool]) -> None:
         for idx, f in enumerate(self.files):
             if defined[idx]:
                 f['attributes'], _ = read_uint32(buffer)
             else:
                 f['attributes'] = None
 
-    def _readTimes(self, fp, files, name):
+    def _readTimes(self, fp: BinaryIO, files: List[Dict[str, Any]], name: str) -> None:
         defined = read_boolean(fp, len(files), checkall=1)
         # NOTE: the "external" flag is currently ignored, should be 0x00
         self.external = fp.read(1)
@@ -720,7 +732,7 @@ class FilesInfo:
             else:
                 files[i][name] = None
 
-    def write(self, file):
+    def write(self, file: BinaryIO):
         assert self.files is not None
         numfiles = len(self.files)
         numemptystreams = 0
@@ -736,7 +748,7 @@ class FilesInfo:
         write_boolean(file, emptystreams, all_defined=False)
         if self.emptyfiles is not None:
             if functools.reduce(or_, self.emptyfiles):  # there are some emptyfile
-                write_byte(Property.EMPTY_FILE)
+                write_byte(file, Property.EMPTY_FILE)
                 write_boolean(file, self.emptyfiles)
         if self.antifiles is not None:
             if functools.reduce(or_, self.antifiles):  # there are some antifiles
@@ -756,7 +768,7 @@ class Header:
     __slot__ = ['solid', 'properties', 'additional_streams', 'main_streams', 'files_info',
                 'size', '_start_pos']
 
-    def __init__(self):
+    def __init__(self) -> None:
         self.solid = False
         self.properties = None
         self.additional_streams = None
@@ -766,17 +778,17 @@ class Header:
         self._start_pos = 0
 
     @classmethod
-    def retrieve(cls, fp, buffer, start_pos):
+    def retrieve(cls, fp: BinaryIO, buffer: BytesIO, start_pos: int):
         obj = cls()
         obj._read(fp, buffer, start_pos)
         return obj
 
-    def _read(self, fp, buffer, start_pos):
+    def _read(self, fp: BinaryIO, buffer: BytesIO, start_pos: int) -> None:
         self._start_pos = start_pos
         fp.seek(self._start_pos)
         self._decode_header(fp, buffer)
 
-    def _decode_header(self, fp, buffer):
+    def _decode_header(self, fp: BinaryIO, buffer: BytesIO) -> None:
         """
         Decode header data or encoded header data from buffer.
         When buffer consist of encoded buffer, it get stream data
@@ -785,16 +797,17 @@ class Header:
         pid = buffer.read(1)
         if not pid:
             # empty archive
-            return None
+            return
         elif pid == Property.HEADER:
-            return self._extract_header_info(buffer)
+            self._extract_header_info(buffer)
+            return
         elif pid != Property.ENCODED_HEADER:
             raise TypeError('Unknown field: %r' % (id))
         # get from encoded header
         streams = StreamsInfo.retrieve(buffer)
-        return self._decode_header(fp, self._get_headerdata_from_streams(fp, streams))
+        self._decode_header(fp, self._get_headerdata_from_streams(fp, streams))
 
-    def _get_headerdata_from_streams(self, fp, streams):
+    def _get_headerdata_from_streams(self, fp: BinaryIO, streams: StreamsInfo) -> BytesIO:
         """get header data from given streams.unpackinfo and packinfo.
         folder data are stored in raw data positioned in afterheader."""
         buffer = io.BytesIO()
@@ -832,7 +845,7 @@ class Header:
         streams.unpackinfo = UnpackInfo()
         streams.unpackinfo.folders = []  # fixme
 
-    def write(self, file, encoded=True):
+    def write(self, file: BinaryIO, encoded: bool = True):
         if encoded:
             self._build_encoded_header()
             write_byte(file, Property.ENCODED_HEADER)
@@ -848,7 +861,7 @@ class Header:
                 self.files_info.write(file)
             write_byte(file, Property.END)
 
-    def _extract_header_info(self, fp):
+    def _extract_header_info(self, fp: BinaryIO) -> None:
         pid = fp.read(1)
         if pid == Property.ARCHIVE_PROPERTIES:
             self.properties = ArchiveProperties.retrieve(fp)
@@ -871,20 +884,20 @@ class SignatureHeader:
 
     __slots__ = ['version', 'startheadercrc', 'nextheaderofs', 'nextheadersize', 'nextheadercrc']
 
-    def __init__(self):
-        self.version = (Configuration.P7ZIP_MAJOR_VERSION, Configuration.P7ZIP_MINOR_VERSION)
-        self.startheadercrc = None
-        self.nextheaderofs = None
-        self.nextheadersize = None
-        self.nextheadercrc = None
+    def __init__(self) -> None:
+        self.version = (Configuration.P7ZIP_MAJOR_VERSION, Configuration.P7ZIP_MINOR_VERSION)  # type: Tuple[bytes, ...]
+        self.startheadercrc = None  # type: Optional[int]
+        self.nextheaderofs = None  # type: Optional[int]
+        self.nextheadersize = None  # type: Optional[int]
+        self.nextheadercrc = None  # type: Optional[int]
 
     @classmethod
-    def retrieve(cls, file):
+    def retrieve(cls, file: BinaryIO):
         obj = cls()
         obj._read(file)
         return obj
 
-    def _read(self, file):
+    def _read(self, file: BinaryIO) -> None:
         file.seek(len(MAGIC_7Z), 0)
         self.version = read_bytes(file, 2)
         self.startheadercrc, _ = read_uint32(file)
@@ -897,7 +910,7 @@ class SignatureHeader:
         if crc != self.startheadercrc:
             raise Bad7zFile('invalid header data')
 
-    def calccrc(self, header):
+    def calccrc(self, header: Header):
         buf = io.BytesIO()
         header.write(buf)
         data = buf.getvalue()
@@ -911,13 +924,15 @@ class SignatureHeader:
         data = buf.getvalue()
         self.startheadercrc = calculate_crc32(data)
 
-    def write(self, file):
+    def write(self, file: BinaryIO):
         assert self.startheadercrc is not None
         assert self.nextheadercrc is not None
+        assert self.nextheaderofs is not None
+        assert self.nextheadersize is not None
         file.seek(0, 0)
         write_bytes(file, MAGIC_7Z)
-        write_byte(file, self.version[0].to_bytes(1, 'little'))
-        write_byte(file, self.version[1].to_bytes(1, 'little'))
+        write_byte(file, self.version[0])
+        write_byte(file, self.version[1])
         write_uint32(file, self.startheadercrc)
         write_uint64(file, self.nextheaderofs)
         write_uint64(file, self.nextheadersize)