+++ py7zr/compression.py
@@ -25,14 +25,15 @@ import bz2
 import concurrent.futures
 import io
 import lzma
-import zlib
+from io import BytesIO
+from typing import Any, BinaryIO, Dict, List, Optional, Union
 
 from py7zr import UnsupportedCompressionMethodError
 from py7zr.helpers import calculate_crc32
 from py7zr.properties import CompressionMethod, Configuration
 
 
-class NullHandler:
+class NullHandler():
     '''Null handler pass to null the data.'''
 
     def __init__(self):
@@ -54,17 +55,17 @@ class NullHandler:
         pass
 
 
-class BufferHandler:
+class BufferHandler():
     '''Buffer handler handles BytesIO/StringIO buffers.'''
 
-    def __init__(self, target):
+    def __init__(self, target: BytesIO) -> None:
         self.buf = target
         self.target = "memory buffer"
 
-    def open(self):
+    def open(self) -> None:
         pass
 
-    def write(self, data):
+    def write(self, data: bytes) -> None:
         self.buf.write(data)
 
     def read(self, size=None):
@@ -76,20 +77,20 @@ class BufferHandler:
     def seek(self, offset, whence=1):
         self.buf.seek(offset, whence)
 
-    def close(self):
+    def close(self) -> None:
         pass
 
 
-class FileHandler:
+class FileHandler():
     '''File handler treat fileish object'''
 
-    def __init__(self, target):
+    def __init__(self, target: str) -> None:
         self.target = target
 
-    def open(self):
+    def open(self) -> None:
         self.fp = open(self.target, 'wb')
 
-    def write(self, data):
+    def write(self, data: bytes) -> None:
         self.fp.write(data)
 
     def read(self, size=None):
@@ -101,23 +102,26 @@ class FileHandler:
     def seek(self, offset, whence=1):
         self.fp.seek(offset, whence)
 
-    def close(self):
+    def close(self) -> None:
         self.fp.close()
+
+
+Handler = Union[NullHandler, BufferHandler, FileHandler]
 
 
 class Worker:
     """Extract worker class to invoke handler"""
 
-    def __init__(self, files, src_start, header):
-        self.target_filepath = {}
+    def __init__(self, files, src_start: int, header) -> None:
+        self.target_filepath = {}  # type: Dict[int, Handler]
         self.files = files
         self.src_start = src_start
         self.header = header
 
-    def set_output_filepath(self, index, func):
+    def set_output_filepath(self, index: int, func: Handler) -> None:
         self.target_filepath[index] = func
 
-    def extract(self, fp, multithread=False):
+    def extract(self, fp: BinaryIO, multithread: bool = False) -> None:
         if multithread:
             numfolders = self.header.main_streams.unpackinfo.numfolders
             positions = self.header.main_streams.packinfo.packpositions
@@ -136,23 +140,20 @@ class Worker:
         else:
             self.extract_single(fp, self.files, self.src_start)
 
-    def extract_single(self, fp, files, src_start):
+    def extract_single(self, fp: BinaryIO, files, src_start: int) -> None:
         fp.seek(src_start)
         for f in files:
+            fileish = self.target_filepath.get(f.id, NullHandler())  # type: Handler
+            fileish.open()
             # Skip empty file read
             if f.emptystream:
-                fileish = self.target_filepath.get(f.id, None)
-                if fileish is not None:
-                    fileish.open()
-                    fileish.write(b'')
-                    fileish.close()
-                continue
-            fileish = self.target_filepath.get(f.id, NullHandler)
-            fileish.open()
-            self.decompress(fp, f.folder, fileish, f.uncompressed[-1], f.compressed)
+                fileish.write(b'')
+            else:
+                self.decompress(fp, f.folder, fileish, f.uncompressed[-1], f.compressed)
             fileish.close()
 
-    def decompress(self, fp, folder, fileish, size, compressed_size):
+    def decompress(self, fp: BinaryIO, folder, fileish: Handler,
+                   size: int, compressed_size: Optional[int]) -> None:
         assert folder is not None
         out_remaining = size
         decompressor = folder.get_decompressor(compressed_size)
@@ -177,11 +178,11 @@ class Worker:
                 print('\nCRC error! expected: {}, real: {}'.format(decompressor.crc, decompressor.digest))
         return
 
-    def archive(self, fp, folder):
+    def archive(self, fp: BinaryIO, folder):
         fp.seek(self.src_start)
         for f in self.files:
             if not f['emptystream']:
-                target = self.target_filepath.get(f.id, None)
+                target = self.target_filepath.get(f.id, NullHandler())  # type: Handler
                 target.open()
                 length = self.compress(fp, folder, target)
                 target.close()
@@ -189,7 +190,7 @@ class Worker:
             self.files.append(f)
         fp.flush()
 
-    def compress(self, fp, folder, f):
+    def compress(self, fp: BinaryIO, folder, f: Handler):
         compressor = folder.get_compressor()
         length = 0
         for indata in f.read(Configuration.get('read_blocksize')):
@@ -203,13 +204,15 @@ class Worker:
         fp.write(arcdata)
         return length
 
-    def register_filelike(self, id, fileish):
+    def register_filelike(self, id: int, fileish: Union[BinaryIO, str, None]) -> None:
         if fileish is None:
             self.set_output_filepath(id, NullHandler())
         elif isinstance(fileish, io.BytesIO):
             self.set_output_filepath(id, BufferHandler(fileish))
-        else:
+        elif isinstance(fileish, str):
             self.set_output_filepath(id, FileHandler(fileish))
+        else:
+            raise
 
 
 class SevenZipDecompressor:
@@ -230,18 +233,51 @@ class SevenZipDecompressor:
     FILTER_ZIP = 0x32
     alt_methods_map = {
         CompressionMethod.MISC_BZIP2: FILTER_BZIP2,
-        CompressionMethod.MISC_ZIP: FILTER_ZIP,
     }
 
+    def __init__(self, coders: List[Dict[str, Any]], size: int, crc: Optional[int]) -> None:
+        self.input_size = size
+        self.consumed = 0  # type: int
+        self.crc = crc
+        self.digest = None  # type: Optional[int]
+        filters = []  # type: List[Dict[str, Any]]
+        try:
+            for coder in coders:
+                if coder['numinstreams'] != 1 or coder['numoutstreams'] != 1:
+                    raise UnsupportedCompressionMethodError('Only a simple compression method is currently supported.')
+                filter = self.lzma_methods_map.get(coder['method'], None)
+                if filter is not None:
+                    properties = coder.get('properties', None)
+                    if properties is not None:
+                        filters[:0] = [lzma._decode_filter_properties(filter, properties)]  # type: ignore
+                    else:
+                        filters[:0] = [{'id': filter}]
+                else:
+                    raise UnsupportedCompressionMethodError
+        except UnsupportedCompressionMethodError as e:
+            filter = self.alt_methods_map.get(coders[0]['method'], None)
+            if len(coders) == 1 and filter is not None:
+                if filter == self.FILTER_BZIP2:
+                    self.decompressor = bz2.BZ2Decompressor()  # type: Union[bz2.BZ2Decompressor, lzma.LZMADecompressor]
+                else:
+                    raise e
+                self.can_partial_decompress = False
+            else:
+                raise e
+        else:
+            self.decompressor = lzma.LZMADecompressor(format=lzma.FORMAT_RAW, filters=filters)
+            self.can_partial_decompress = True
+        self.filters = filters
+
     @property
-    def needs_input(self):
+    def needs_input(self) -> bool:
         return self.decompressor.needs_input
 
     @property
-    def eof(self):
+    def eof(self) -> bool:
         return self.decompressor.eof
 
-    def decompress(self, data, max_length=None):
+    def decompress(self, data: bytes, max_length: Optional[int] = None) -> bytes:
         self.consumed += len(data)
         if max_length is not None:
             folder_data = self.decompressor.decompress(data, max_length=max_length)
@@ -257,48 +293,11 @@ class SevenZipDecompressor:
         return self.decompressor.unused_data
 
     @property
-    def remaining_size(self):
+    def remaining_size(self) -> int:
         return self.input_size - self.consumed
 
     def check_crc(self):
         return self.crc == self.digest
-
-    def __init__(self, coders, size, crc):
-        self.decompressor = None
-        self.input_size = size
-        self.consumed = 0
-        self.crc = crc
-        self.digest = None
-        filters = []
-        try:
-            for coder in coders:
-                if coder['numinstreams'] != 1 or coder['numoutstreams'] != 1:
-                    raise UnsupportedCompressionMethodError('Only a simple compression method is currently supported.')
-                filter = self.lzma_methods_map.get(coder['method'], None)
-                if filter is not None:
-                    properties = coder.get('properties', None)
-                    if properties is not None:
-                        filters[:0] = [lzma._decode_filter_properties(filter, properties)]
-                    else:
-                        filters[:0] = [{'id': filter}]
-                else:
-                    raise UnsupportedCompressionMethodError
-        except UnsupportedCompressionMethodError as e:
-            filter = self.alt_methods_map.get(coders[0]['method'], None)
-            if len(coders) == 1 and filter is not None:
-                if filter == self.FILTER_BZIP2:
-                    self.decompressor = bz2.BZ2Decompressor()
-                elif filter == self.FILTER_ZIP:
-                    self.decompressor = zlib.decompressobj(-15)
-                else:
-                    raise e
-                self.can_partial_decompress = False
-            else:
-                raise e
-        else:
-            self.decompressor = lzma.LZMADecompressor(format=lzma.FORMAT_RAW, filters=filters)
-            self.can_partial_decompress = True
-        self.filters = filters
 
 
 class SevenZipCompressor():
@@ -330,7 +329,7 @@ class SevenZipCompressor():
         return self.compressor.flush()
 
 
-def get_methods_names(coders):
+def get_methods_names(coders: List[dict]) -> List[str]:
     methods_name_map = {
         CompressionMethod.LZMA2: "LZMA2",
         CompressionMethod.LZMA: "LZMA",