+++ py7zr/properties.py
@@ -32,13 +32,13 @@ MAGIC_7Z = binascii.unhexlify('377abcaf271c')
 class Configuration:
     '''Singleton global configuration holder'''
 
-    P7ZIP_MAJOR_VERSION = 0
-    P7ZIP_MINOR_VERSION = 4
+    P7ZIP_MAJOR_VERSION = b'\x00'
+    P7ZIP_MINOR_VERSION = b'\x04'
     read_blocksize = 32248
     _instance = None
 
     @classmethod
-    def get(cls, key):
+    def get(cls, key: str) -> int:
         if cls._instance is None:
             cls._instance = cls()
         return getattr(cls._instance, key)